feat: Make profile page use SSR
